refactor: create or update email thread outside of `sendEmail`
